+++ dune/stuff/common/ranges.hh
@@ -71,7 +71,7 @@ namespace Common {
 
 
 //! adapter enabling view usage in range-based for
-template <class GridPartViewType, int codim = 0>
+template <class GridPartViewType, size_t codim = 0>
 class EntityRange
 {
   const GridPartViewType& part_view_;
@@ -95,7 +95,7 @@ public:
 }; // class EntityRange
 
 
-template <class GridPartViewType, int codim = 0>
+template <class GridPartViewType, size_t codim = 0>
 class DUNE_DEPRECATED_MSG("Use EntityRange instead (22.09.2014)!") ViewRange
     : public EntityRange<GridPartViewType, codim>
 {
@@ -108,14 +108,14 @@ public:
 }; // class ViewRange
 
 
-template <class GridViewTraits, int codim = 0>
+template <class GridViewTraits, size_t codim = 0>
 EntityRange<Dune::GridView<GridViewTraits>, codim> DUNE_DEPRECATED_MSG("Use entityRange instead (22.09.2014)!")
     viewRange(const Dune::GridView<GridViewTraits>& view)
 {
   return EntityRange<Dune::GridView<GridViewTraits>, codim>(view);
 }
 
-template <class GridViewTraits, int codim = 0>
+template <class GridViewTraits, size_t codim = 0>
 EntityRange<Dune::GridView<GridViewTraits>, codim> entityRange(const Dune::GridView<GridViewTraits>& view)
 {
   return EntityRange<Dune::GridView<GridViewTraits>, codim>(view);
@@ -125,14 +125,14 @@ EntityRange<Dune::GridView<GridViewTraits>, codim> entityRange(const Dune::GridV
 #if HAVE_DUNE_FEM
 
 
-template <class GP, int codim = 0>
+template <class GP, size_t codim = 0>
 EntityRange<Dune::Fem::GridPartInterface<GP>, codim> DUNE_DEPRECATED_MSG("Use entityRange instead (22.09.2014)!")
     viewRange(const Dune::Fem::GridPartInterface<GP>& part)
 {
   return EntityRange<Dune::Fem::GridPartInterface<GP>, codim>(part);
 }
 
-template <class GP, int codim = 0>
+template <class GP, size_t codim = 0>
 EntityRange<Dune::Fem::GridPartInterface<GP>, codim> entityRange(const Dune::Fem::GridPartInterface<GP>& part)
 {
   return EntityRange<Dune::Fem::GridPartInterface<GP>, codim>(part);
@@ -300,7 +300,7 @@ public:
   }
 };
 
-template <int codim, class DiscreteFunctionspaceType, class EntityType>
+template <size_t codim, class DiscreteFunctionspaceType, class EntityType>
 LagrangePointSetRange<typename DiscreteFunctionspaceType::GridPartType, DiscreteFunctionspaceType::polynomialOrder,
                       codim>
 lagrangePointSetRange(const DiscreteFunctionspaceType& space, const EntityType& entity, const size_t subEntity)
@@ -310,9 +310,9 @@ lagrangePointSetRange(const DiscreteFunctionspaceType& space, const EntityType& 
                                codim>(space, entity, subEntity);
 }
 
-template <class LgPointSetType, int codim = 1>
+template <class LgPointSetType, size_t codim = 1>
 LagrangePointSetRange<typename LgPointSetType::GridPartType, LgPointSetType::polynomialOrder, codim>
-lagrangePointSetRange(const LgPointSetType& lpset, const int subEntity)
+lagrangePointSetRange(const LgPointSetType& lpset, const size_t subEntity)
 {
   return LagrangePointSetRange<typename LgPointSetType::GridPartType, LgPointSetType::polynomialOrder, codim>(
       lpset, subEntity);
